SK-Q01: Rename globals.scss to globals.css and update imports; refactor form components in Login, Register, ForgotPassword, and MailSent pages
