+++ src/app/(auth)/mail-sent/page.tsx
@@ -1,9 +1,8 @@
-import Input from "@/components/Form/Input";
-import Button from "@/components/Form/Button";
+import {Button} from "@/components/ui/button";
 
 function MailSent() {
     return (
-        <div className="bg-white p-12 rounded-lg shadow-lg max-w-xl">
+        <div className="bg-white p-12 rounded-lg shadow-lg w-xl  max-w-xl">
             <form>
                 <h1 className="text-3xl font-bold bg-gradient-to-r from-[#FFFF42] to-[#ED2409] bg-clip-text text-transparent text-center mb-12 mt-2">
                     Đặt lại mật khẩu
@@ -12,7 +11,9 @@ function MailSent() {
                 <Button
                     type="button"
                     name="Đi đến mail"
-                />
+                    size="lg"
+                    className="w-full mt-4 bg-orange-500"
+                >Đi đến mail</Button>
             </form>
         </div>
     );
