+++ src/app/(auth)/reset-password/page.tsx
@@ -1,9 +1,9 @@
-import Input from "@/components/Form/Input";
-import Button from "@/components/Form/Button";
+import {Input} from "@/components/ui/input";
+import {Button} from "@/components/ui/button";
 
 function ForgotPassword() {
     return (
-        <div className="bg-white p-12 rounded-lg shadow-lg max-w-xl">
+        <div className="bg-white p-12 rounded-lg shadow-lg w-xl  max-w-xl">
             <form>
                 <h1 className="text-3xl font-bold bg-gradient-to-r from-[#FFFF42] to-[#ED2409] bg-clip-text text-transparent text-center mb-12 mt-2">
                     Quên mật khẩu
@@ -12,11 +12,13 @@ function ForgotPassword() {
                     type="email"
                     placeholder="Email*"
                     name="email"
+                    className="h-12"
                 />
                 <Button
                     type="submit"
-                    name="Lấy lại mật khẩu"
-                />
+                    size="lg"
+                    className="w-full mt-4 bg-orange-500"
+                >Lấy lại mật khẩu</Button>
             </form>
         </div>
     );
