+++ src/app/globals.css
@@ -1,5 +1,5 @@
 @import "tailwindcss";
-@import "tw-animate-css";
+@import 'tw-animate-css';
 
 @custom-variant dark (&:is(.dark *));
